feat: added next config wrapper utility
