+++ src/cacheHandler/index.ts
@@ -5,13 +5,14 @@ import { S3Cache } from './strategy/s3'
 
 const { serverRuntimeConfig } = getConfig() || {}
 const config: CacheConfig | undefined = serverRuntimeConfig?.nextServerlessCacheConfig
+const staticBucketName = serverRuntimeConfig?.staticBucketName || ''
 
 Cache.setConfig({
   cacheCookies: config?.cacheCookies ?? [],
   cacheQueries: config?.cacheQueries ?? [],
   noCacheMatchers: config?.noCacheRoutes ?? [],
   enableDeviceSplit: config?.enableDeviceSplit,
-  cache: new S3Cache(process.env.STATIC_BUCKET_NAME!)
+  cache: new S3Cache(staticBucketName)
 })
 
 export default Cache
